(store) fix accidental deletion of the primary database (#578)
The IPC removes the db_path dir to remove the secondary droppings but
this shouldn't happen on a primary database. Only remove non-primary
stores when handling Unix Domain handle_conn calls.

+++ src/ipc.rs
@@ -77,8 +77,10 @@ impl IpcActor {
                                     },
                                     Ok(_) => { info!("handled connection"); },
                                 };
-                                debug!("Removing readonly instance at {}", store.db_path.clone().display());
-                                tokio::fs::remove_dir_all(&store.db_path).await.ok();
+                                if !store.is_primary {
+                                    debug!("Removing readonly instance at {}", store.db_path.clone().display());
+                                    tokio::fs::remove_dir_all(&store.db_path).await.ok();
+                                }
                             });
                         }
                     }
